refactor(Posts): seperate Posts & PostsPage
