+++ src/client/components/posts/index.ts
@@ -1,3 +1,8 @@
+import CommentCard from './Comment';
+import CommentBox from './CommentBox';
 import CreateOrEditPost from './CreateOrEditPost';
+import Post from './Post';
+import Posts from './Posts';
+import Votes from './Votes';
 
-export { CreateOrEditPost as NewPost };
+export { CreateOrEditPost, CommentCard, CommentBox, Posts, Post, Votes };
