+++ src/server/handlers/posts/servePosts.ts
@@ -19,7 +19,6 @@ export const servePosts: RequestHandler = async (req, res, _): Promise<void> => 
 
     res.status(200).send(clientReadyPosts);
   } catch (err) {
-    console.log('servePosts err:', err);
-    res.status(200).send(err);
+    res.status(500).send(err);
   }
 };
